requirements.txt
langchain_community==0.2.0
langchain_openai==0.1.7
langgraph==0.0.55
wikipedia==1.4.0
scikit-learn==1.5.0
langchain_fireworks==0.1.3
python-dotenv==1.0.1
duckduckgo-search==6.1.0
pygraphviz==1.13
markdown2==2.4.13
weasyprint==62.1
streamlit==1.35.0